requirements.txt
openai
requests
python-dotenv
fastapi
uvicorn
psycopg2-binary